+++ src/schema/schema-generator.ts
@@ -145,19 +145,20 @@ export abstract class SchemaGenerator {
 
   private static buildTypesInfo() {
     this.unionTypesInfo = getMetadataStorage().unions.map<UnionTypeInfo>(unionMetadata => {
+      const unionClassTypes = unionMetadata.getClassTypes();
       return {
         unionSymbol: unionMetadata.symbol,
         type: new GraphQLUnionType({
           name: unionMetadata.name,
           description: unionMetadata.description,
           types: () =>
-            unionMetadata.classTypes.map(
+            unionClassTypes.map(
               objectType => this.objectTypesInfo.find(type => type.target === objectType)!.type,
             ),
           resolveType: unionMetadata.resolveType
             ? this.getResolveTypeFunction(unionMetadata.resolveType)
             : instance => {
-                const instanceTarget = unionMetadata.classTypes.find(
+                const instanceTarget = unionClassTypes.find(
                   ClassType => instance instanceof ClassType,
                 );
                 if (!instanceTarget) {